feat(i18n): refactored components to use i18n
